+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Keep a changelog!
+- Add --verbose option to emit more verbose logs. Otherwise will emit only changes to configuration
+
+### Fixed
+- Fix handling of changelog-add strings :D
 
 ## [0.1.0] - 2023-05-22
 ### Added